Part B and C
This is the code I did for assigning player abilities and changes based on the terrain on which the player is on. I'm not sure how you did part A, if you put the players into an array or how you did the slots so the code obviously can't run without that part. There could be mistakes in the code because I can't compile without Part A but feel free to change anything you feel needs changing.